test(NetworkClient): restructure dependencies
