--- MovieQuizTests/MockURLProtocol.swift
@@ -1,37 +1,0 @@
-import Foundation
-
-class MockURLProtocol: URLProtocol {
-    typealias MockResponse = (HTTPURLResponse, Data?)
-    static var requestHandler: ((URLRequest) throws -> MockResponse)?
-
-    override class func canInit(with request: URLRequest) -> Bool {
-        true
-    }
-
-    override class func canonicalRequest(for request: URLRequest) -> URLRequest {
-        request
-    }
-
-    override func startLoading() {
-        guard let handler = MockURLProtocol.requestHandler else {
-            fatalError("Handler is unavailable.")
-        }
-
-        do {
-            let (response, data) = try handler(request)
-            client?.urlProtocol(
-                self, didReceive: response, cacheStoragePolicy: .notAllowed
-            )
-
-            if let data = data {
-                client?.urlProtocol(self, didLoad: data)
-            }
-
-            client?.urlProtocolDidFinishLoading(self)
-        } catch {
-            client?.urlProtocol(self, didFailWithError: error)
-        }
-    }
-
-    override func stopLoading() { }
-}
