+++ MovieQuizTests/NetworkClientTests.swift
@@ -1,6 +1,43 @@
 import XCTest
 @testable import MovieQuiz
 
+class StubURLProtocol: URLProtocol {
+    typealias MockResponse = (HTTPURLResponse, Data?)
+    static var requestHandler: ((URLRequest) throws -> MockResponse)?
+
+    override class func canInit(with request: URLRequest) -> Bool {
+        true
+    }
+
+    override class func canonicalRequest(for request: URLRequest) -> URLRequest {
+        request
+    }
+
+    override func startLoading() {
+        guard let handler = StubURLProtocol.requestHandler else {
+            fatalError("Handler is unavailable.")
+        }
+
+        do {
+            let (response, data) = try handler(request)
+            client?.urlProtocol(
+                self, didReceive: response, cacheStoragePolicy: .notAllowed
+            )
+
+            if let data = data {
+                client?.urlProtocol(self, didLoad: data)
+            }
+
+            client?.urlProtocolDidFinishLoading(self)
+        } catch {
+            client?.urlProtocol(self, didFailWithError: error)
+        }
+    }
+
+    override func stopLoading() { }
+}
+
+
 // swiftlint:disable force_unwrapping implicitly_unwrapped_optional
 final class NetworkClientTests: XCTestCase {
     var urlSession: URLSession!
@@ -8,7 +45,7 @@ final class NetworkClientTests: XCTestCase {
     override func setUp() {
         super.setUp()
         let configuration: URLSessionConfiguration = .ephemeral
-        configuration.protocolClasses = [MockURLProtocol.self]
+        configuration.protocolClasses = [StubURLProtocol.self]
         urlSession = URLSession(configuration: configuration)
     }
 
@@ -18,7 +55,7 @@ final class NetworkClientTests: XCTestCase {
 
         let url = URL.apiURL(.mostPopularMovies)
         let urlData = "[]".data(using: .utf8)!
-        MockURLProtocol.requestHandler = { _ in
+        StubURLProtocol.requestHandler = { _ in
             let response = HTTPURLResponse(
                 url: url,
                 statusCode: 200,
@@ -50,7 +87,7 @@ final class NetworkClientTests: XCTestCase {
         let sut = NetworkClient(urlSession: urlSession)
         let url = URL.apiURL(.mostPopularMovies)
 
-        MockURLProtocol.requestHandler = { _ in
+        StubURLProtocol.requestHandler = { _ in
             let response = HTTPURLResponse(
                 url: url,
                 statusCode: 500,
@@ -82,7 +119,7 @@ final class NetworkClientTests: XCTestCase {
         let url = URL.apiURL(.mostPopularMovies)
         let error = NSError(domain: "test", code: 100)
 
-        MockURLProtocol.requestHandler = { _ in
+        StubURLProtocol.requestHandler = { _ in
             throw error
         }
